fixed skill action taking ticks as casttime and not seconds

+++ src/main/java/de/raidcraft/skills/api/ability/Ability.java
@@ -48,7 +48,7 @@ public interface Ability<T extends CharacterTemplate> {
 
     int getTotalDamage();
 
-    int getTotalCastTime();
+    double getTotalCastTime();
 
     int getTotalRange();
